+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 12.12.0-beta.2 (Unreleased)
 - Added support for the 2020-10-02 service version.
+- Added support to specify Parquet Input Serialization when querying a blob.
 
 ## 12.12.0-beta.1 (2021-05-13)
 - Added support for the 2020-08-04 service version.